backend/requirements.txt
python-dotenv
asyncpg
fastapi
uvicorn
orjson